working on resolving the error that says I have a double id value

+++ public/javascript/delete-post.js
@@ -2,35 +2,42 @@ async function deleteFormHandler(event) {
 
     event.preventDefault();
 
-    const id = window.location.toString().split('/')[
+    const title = document.querySelector('input[name="post-title"]').value.trim();
+
+    const post_id = window.location.toString().split('/')[
         window.location.toString().split('/')[
             window.location.toString().split('/').length-1
         ]
     ];
 
-    const response = await fetch('/api/posts/:id', {
+    if (title) {
+        const response = await fetch('/api/posts/:id', {
 
-        method: 'DELETE',
+            method: 'DELETE',
+    
+            body: JSON.stringify({
+                
+                post_id,
 
-        body: JSON.stringify({
-            
-            post_id: id
+                title
+    
+            }),
+    
+            headers: {
+    
+                'Content-Type': 'application/json'
+    
+            }
 
-        }),
-
-        headers: {
-
-            'Content-Type': 'application/json'
-
+        });
+    
+        if (response.ok) {
+            document.location.replace('/dashboard/')
+        } else {
+            alert(response.statusText)
         }
-    });
-
-    if (response.ok) {
-        document.location.replace('/dashboard')
-    } else {
-        alert(response.statusText)
+    
     }
-
 }
 
 document.querySelector('.delete-post-btn').addEventListener('click', deleteFormHandler);
